update model download template

+++ job-template/job/model_download/launcher.py
@@ -22,6 +22,7 @@ def download(**kwargs):
         'Authorization': KFJ_CREATOR
     }
     model_path=""
+    exist_model = {}
     # 从注册的模型中下载模型
     if kwargs['from']=='模型管理':
         url = host + "/training_model_modelview/api/?form_data=" + json.dumps({
@@ -36,7 +37,8 @@ def download(**kwargs):
                     "opr": "eq",
                     "value": kwargs['model_version']
                 }
-            ]
+            ],
+            "columns":['id','project','name','version','describe','path','framework','run_id','run_time','metrics','md5','api_type','pipeline_id']
         })
 
         # print(url)
@@ -49,6 +51,10 @@ def download(**kwargs):
                 print(exist_model)
                 if exist_model['path']:
                     model_path = exist_model['path']
+        else:
+            print('访问平台获取模型失败')
+            print(res.content)
+            exit(1)
 
     elif kwargs['from']=='推理服务':
         filters = [
@@ -72,7 +78,9 @@ def download(**kwargs):
 
 
         url = host+"/inferenceservice_modelview/api/?form_data="+json.dumps({
-            "filters":filters
+            "filters":filters,
+            "columns":  ['service_type','project', 'name', 'label','model_name', 'model_version', 'images', 'model_path', 'images', 'volume_mount','sidecar','working_dir', 'command', 'env', 'resource_memory',
+                    'resource_cpu', 'resource_gpu', 'min_replicas', 'max_replicas', 'ports', 'inference_host_url','hpa','priority', 'canary', 'shadow', 'health','model_status','expand','metrics','deploy_history','host','inference_config']
         })
 
         # print(url)
@@ -85,10 +93,18 @@ def download(**kwargs):
                 print(exist_service)
                 if exist_service['model_path']:
                     model_path = exist_service['model_path']
+        else:
+            print('访问平台获取模型失败')
+            print(res.content)
+            exit(1)
 
-    if model_path:
-        os.makedirs(model_path,exist_ok=True)
+    if model_path and os.path.exists(model_path):
+        os.makedirs(kwargs['save_path'],exist_ok=True)
         shutil.copy2(model_path,kwargs['save_path'])
+        # 同时将模型信息写入到存储中,比如计算指标
+        if kwargs['from']=='模型管理':
+            if exist_model:
+                json.dump(exist_model,open(os.path.join(kwargs['save_path'],f'{exist_model["name"]}.{exist_model["version"]}.json')))
     else:
         print('未发现模型')
         exit(1)
